// src/shared/api/models/getApiInsightsTaxReportsType200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsTaxReportsType200Data } from './getApiInsightsTaxReportsType200Data';

export type GetApiInsightsTaxReportsType200 = {
    data: GetApiInsightsTaxReportsType200Data;
    message: string;
};