// src/shared/api/models/getApiInsightsTaxReportsTypeTaxType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiInsightsTaxReportsTypeTaxType =
    (typeof GetApiInsightsTaxReportsTypeTaxType)[keyof typeof GetApiInsightsTaxReportsTypeTaxType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiInsightsTaxReportsTypeTaxType = {
    tax: 'tax',
    notax: 'notax',
} as const;